+++ CM7/Src/page_audio.c
@@ -114,7 +114,7 @@ uint8_t page_init_audio(void) {
 	button00.y_pos = 70;
 	button00.width = 160;
 	button00.height = 42;
-	button00.icon = (uint32_t)ICON_32_Audio;
+	button00.icon = ICON_32_Audio;
 	button00.text = "Left";
 	button00.state = GUI_STATE_ENABLED;
 	BSP_LCD_TP_RegisterArea(10, 240, 70, 160, 42, button00_callback);
@@ -124,7 +124,7 @@ uint8_t page_init_audio(void) {
 	button01.y_pos = 70;
 	button01.width = 160;
 	button01.height = 42;
-	button01.icon = (uint32_t)ICON_32_Audio;
+	button01.icon = ICON_32_Audio;
 	button01.text = "Stereo";
 	button01.state = GUI_STATE_ENABLED;
 	BSP_LCD_TP_RegisterArea(11, 425, 70, 160, 42, button01_callback);
@@ -134,7 +134,7 @@ uint8_t page_init_audio(void) {
 	button02.y_pos = 70;
 	button02.width = 160;
 	button02.height = 42;
-	button02.icon = (uint32_t)ICON_32_Audio;
+	button02.icon = ICON_32_Audio;
 	button02.text = "Right";
 	button02.state = GUI_STATE_ENABLED;
 	BSP_LCD_TP_RegisterArea(12, 610, 70, 160, 42, button02_callback);
@@ -153,7 +153,7 @@ uint8_t page_init_audio(void) {
 	button03.y_pos = 260;
 	button03.width = 160;
 	button03.height = 42;
-	button03.icon = (uint32_t)ICON_32_Tools;
+	button03.icon = ICON_32_Tools;
 	button03.text = "Save";
 	button03.state = GUI_STATE_ENABLED;
 	BSP_LCD_TP_RegisterArea(14, 590, 260, 160, 42, button03_callback);
